+++ benchmark/run_1yr_tt_benchmark.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-'''
+"""
 run_1yr_benchmark.py: Driver script for creating benchmark plots
 for the 1-year TransportTracers benchmark.
 
@@ -33,14 +33,14 @@ Remarks:
     need to use these commands to disable the X-window functionality.
 
         import os
-        os.environ['QT_QPA_PLATFORM']='offscreen'
+        os.environ["QT_QPA_PLATFORM"]="offscreen"
 
     For more information, please see this issue posted at the ipython site:
 
         https://github.com/ipython/ipython/issues/10627
 
     This issue might be fixed in matplotlib 3.0.
-'''
+"""
 
 # =====================================================================
 # Imports and global settings (you should not need to edit these)
@@ -60,11 +60,11 @@ import numpy as np
 import warnings
 
 # Tell matplotlib not to look for an X-window
-os.environ['QT_QPA_PLATFORM']='offscreen'
+os.environ["QT_QPA_PLATFORM"]="offscreen"
 
 # Suppress harmless run-time warnings (mostly about underflow in division)
-warnings.filterwarnings('ignore', category=RuntimeWarning)
-warnings.filterwarnings('ignore', category=UserWarning)
+warnings.filterwarnings("ignore", category=RuntimeWarning)
+warnings.filterwarnings("ignore", category=UserWarning)
 
 # =====================================================================
 # Configurables
@@ -74,11 +74,11 @@ warnings.filterwarnings('ignore', category=UserWarning)
 # Note: When doing GCHP vs GCC comparisions gchp_dev will be compared to
 #  gcc_dev (not gcc_ref!). This ensures consistency in version names when
 #  doing GCHP vs GCC diff-of-diffs (mps, 6/27/19)
-maindir  = '/path/to/main/directory'
-gcc_ref_version = 'gcc_ref_version_string'
-gcc_dev_version = 'gcc_dev_version_string'
-gchp_ref_version = 'gchp_ref_version_string'
-gchp_dev_version = 'gchp_dev_version_string'
+maindir  = "/path/to/main/directory"
+gcc_ref_version = "gcc_ref_version_string"
+gcc_dev_version = "gcc_dev_version_string"
+gchp_ref_version = "gchp_ref_version_string"
+gchp_dev_version = "gchp_dev_version_string"
 
 # Comparisons to run (edit as needed)
 # NOTE: For now, just develop GCC vs GCC and add others later (bmy, 10/11/19)
@@ -90,7 +90,7 @@ gchp_vs_gcc_diff_of_diffs = False
 ########################################################################
 #### TransportTracers BENCHMARK OPTIONS                             ####
 ########################################################################
-bmk_type     = 'TransportTracersBenchmark'
+bmk_type     = "TransportTracersBenchmark"
 plot_conc    = True
 plot_wetdep  = True
 budget_table = True
@@ -113,41 +113,41 @@ bmk_start = np.datetime64(bmk_start_str)
 bmk_end = np.datetime64(bmk_end_str)
 
 # Data directories (edit as needed)
-gcc_vs_gcc_refdir   = join(maindir, gcc_ref_version,  'OutputDir')
-gcc_vs_gcc_devdir   = join(maindir, gcc_dev_version,  'OutputDir')
-gchp_vs_gcc_refdir  = join(maindir, gcc_dev_version,  'OutputDir')
-gchp_vs_gcc_devdir  = join(maindir, gchp_dev_version, 'OutputDir')
-gchp_vs_gchp_refdir = join(maindir, gchp_ref_version, 'OutputDir')
-gchp_vs_gchp_devdir = join(maindir, gchp_dev_version, 'OutputDir')
+gcc_vs_gcc_refdir   = join(maindir, gcc_ref_version,  "OutputDir")
+gcc_vs_gcc_devdir   = join(maindir, gcc_dev_version,  "OutputDir")
+gchp_vs_gcc_refdir  = join(maindir, gcc_dev_version,  "OutputDir")
+gchp_vs_gcc_devdir  = join(maindir, gchp_dev_version, "OutputDir")
+gchp_vs_gchp_refdir = join(maindir, gchp_ref_version, "OutputDir")
+gchp_vs_gchp_devdir = join(maindir, gchp_dev_version, "OutputDir")
 
 # Restart file directories (edit as needed)
-gcc_vs_gcc_refrstdir   = join(maindir, gcc_ref_version,  'restarts')
-gcc_vs_gcc_devrstdir   = join(maindir, gcc_dev_version,  'restarts')
-gchp_vs_gcc_refrstdir  = join(maindir, gcc_dev_version,  'restarts')
-gchp_vs_gcc_devrstdir  = join(maindir, gchp_dev_version, 'restarts')
-gchp_vs_gchp_refrstdir = join(maindir, gchp_ref_version, 'restarts')
-gchp_vs_gchp_devrstdir = join(maindir, gchp_dev_version, 'restarts')
+gcc_vs_gcc_refrstdir   = join(maindir, gcc_ref_version,  "restarts")
+gcc_vs_gcc_devrstdir   = join(maindir, gcc_dev_version,  "restarts")
+gchp_vs_gcc_refrstdir  = join(maindir, gcc_dev_version,  "restarts")
+gchp_vs_gcc_devrstdir  = join(maindir, gchp_dev_version, "restarts")
+gchp_vs_gchp_refrstdir = join(maindir, gchp_ref_version, "restarts")
+gchp_vs_gchp_devrstdir = join(maindir, gchp_dev_version, "restarts")
 
 # Plots directories (edit as needed)
-gcc_vs_gcc_plotsdir    = join(maindir, gcc_dev_version, 'Plots')
+gcc_vs_gcc_plotsdir    = join(maindir, gcc_dev_version, "Plots")
 gchp_vs_gchp_plotsdir  = join(maindir, gchp_dev_version,
-                              'Plots/GCHP_version_comparison')
+                              "Plots/GCHP_version_comparison")
 gchp_vs_gcc_plotsdir   = join(maindir, gchp_dev_version,
-                              'Plots/GCHP_GCC_comparison')
+                              "Plots/GCHP_GCC_comparison")
 diff_of_diffs_plotsdir = join(maindir, gchp_dev_version,
-                              'Plots/GCHP_GCC_diff_of_diffs')
+                              "Plots/GCHP_GCC_diff_of_diffs")
 
 # Plot title strings (edit as needed)
 # For gchp_vs_gcc_refstr use gcc_dev_version, not ref (mps, 6/27/19)
-gcc_vs_gcc_refstr    = '{}'.format(gcc_ref_version)
-gcc_vs_gcc_devstr    = '{}'.format(gcc_dev_version)
-gchp_vs_gcc_refstr   = '{}'.format(gcc_dev_version)
-gchp_vs_gcc_devstr   = '{}'.format(gchp_dev_version)
-gchp_vs_gchp_refstr  = '{}'.format(gchp_ref_version)
-gchp_vs_gchp_devstr  = '{}'.format(gchp_dev_version)
-diff_of_diffs_refstr = '{} - {}'.format(gcc_dev_version,
+gcc_vs_gcc_refstr    = "{}".format(gcc_ref_version)
+gcc_vs_gcc_devstr    = "{}".format(gcc_dev_version)
+gchp_vs_gcc_refstr   = "{}".format(gcc_dev_version)
+gchp_vs_gcc_devstr   = "{}".format(gchp_dev_version)
+gchp_vs_gchp_refstr  = "{}".format(gchp_ref_version)
+gchp_vs_gchp_devstr  = "{}".format(gchp_dev_version)
+diff_of_diffs_refstr = "{} - {}".format(gcc_dev_version,
                                         gcc_ref_version)
-diff_of_diffs_devstr = '{} - {}'.format(gchp_dev_version,
+diff_of_diffs_devstr = "{} - {}".format(gchp_dev_version,
                                         gchp_ref_version)
 
 # =====================================================================
@@ -157,29 +157,29 @@ diff_of_diffs_devstr = '{} - {}'.format(gchp_dev_version,
 ###############################################################################
 # Under development, comment these out for now
 ## Paths to concentration after chemistry files
-#gcc_vs_gcc_refcac = core.get_gcc_filepath(gcc_vs_gcc_refdir, 'ConcAfterChem',
+#gcc_vs_gcc_refcac = core.get_gcc_filepath(gcc_vs_gcc_refdir, "ConcAfterChem",
 #                                          year=bmk_year, month=bmk_months)
 #
-#gcc_vs_gcc_devcac = core.get_gcc_filepath(gcc_vs_gcc_devdir, 'ConcAfterChem',
+#gcc_vs_gcc_devcac = core.get_gcc_filepath(gcc_vs_gcc_devdir, "ConcAfterChem",
 #                                          year=bmk_year, month=bmk_months)
 ###############################################################################
 
 # Monthly array of dates
-bmk_delta_1m = np.timedelta64(1, 'M')
+bmk_delta_1m = np.timedelta64(1, "M")
 bmk_months = np.arange(bmk_start, bmk_end,
-                       step=bmk_delta_1m, dtype='datetime64[M]')
+                       step=bmk_delta_1m, dtype="datetime64[M]")
 
 # Get the benchmark year from the datetime64 object
-bmk_year = bmk_months[0].astype('datetime64[Y]').astype(int) + 1970
+bmk_year = bmk_months[0].astype("datetime64[Y]").astype(int) + 1970
 
 # Seasonal array of dates
-bmk_delta_3m = np.timedelta64(3, 'M')
+bmk_delta_3m = np.timedelta64(3, "M")
 bmk_seasons = np.arange(bmk_start, bmk_end,
-                        step=bmk_delta_3m, dtype='datetime64[M]')
+                        step=bmk_delta_3m, dtype="datetime64[M]")
 bmk_nseasons = len(bmk_seasons)
 
 # Names for each season (e.g. Jan2016, Apr2016, Jul2016, Oct2016)
-bmk_seasons_names = ['Jan', 'Apr', 'Jul', 'Oct']
+bmk_seasons_names = ["Jan", "Apr", "Jul", "Oct"]
 bmk_seasons_names = [v + str(bmk_year) for v in bmk_seasons_names]
 
 # Seconds in the benchmark year
@@ -196,7 +196,7 @@ if not os.path.isdir(sigdiff_dir):
     os.mkdir(sigdiff_dir)
 gcc_vs_gcc_sigdiff = {}
 for mon_yr_str in bmk_seasons_names:
-    vstr = '{}_vs_{}.{}'.format(
+    vstr = "{}_vs_{}.{}".format(
         gcc_vs_gcc_refstr, gcc_vs_gcc_devstr, mon_yr_str)
 
     sigdiff_files = [
@@ -211,19 +211,19 @@ for mon_yr_str in bmk_seasons_names:
 ##############################################################################
 # For now we are only developing the GCC vs. GCC 1-yr benchmark code
 # so leave this commented out (bmy, 10/15/19)
-#vstr = '{}_vs_{}'{}.format(gchp_vs_gcc_refstr, gchp_vs_gcc_devstr)
+#vstr = "{}_vs_{}"{}.format(gchp_vs_gcc_refstr, gchp_vs_gcc_devstr)
 #gchp_vs_gcc_sigdiff = [
-#    join(gchp_vs_gcc_plotsdir, '{}_sig_diffs_sfc.txt'.format(vstr)),
-#    join(gchp_vs_gcc_plotsdir, '{}_sig_diffs_500hpa.txt'.format(vstr)),
-#    join(gchp_vs_gcc_plotsdir, '{}_sig_diffs_zonalmean.txt'.format(vstr)),
-#    join(gchp_vs_gcc_plotsdir, '{}_sig_diffs_emissions.txt'.format(vstr))]
+#    join(gchp_vs_gcc_plotsdir, "{}_sig_diffs_sfc.txt".format(vstr)),
+#    join(gchp_vs_gcc_plotsdir, "{}_sig_diffs_500hpa.txt".format(vstr)),
+#    join(gchp_vs_gcc_plotsdir, "{}_sig_diffs_zonalmean.txt".format(vstr)),
+#    join(gchp_vs_gcc_plotsdir, "{}_sig_diffs_emissions.txt".format(vstr))]
 #
-#vstr = '{}_vs_{}'.format(gchp_vs_gchp_refstr, gchp_vs_gchp_devstr)
+#vstr = "{}_vs_{}".format(gchp_vs_gchp_refstr, gchp_vs_gchp_devstr)
 #gchp_vs_gchp_sigdiff = [
-#    join(gchp_vs_gchp_plotsdir, '{}_sig_diffs_sfc.txt'.format(vstr)),
-#    join(gchp_vs_gchp_plotsdir, '{}_sig_diffs_500hpa.txt'.format(vstr)),
-#    join(gchp_vs_gchp_plotsdir, '{}_sig_diffs_zonalmean.txt'.format(vstr)),
-#    join(gchp_vs_gchp_plotsdir, '{}_sig_diffs_emissions.txt').format(vstr)]
+#    join(gchp_vs_gchp_plotsdir, "{}_sig_diffs_sfc.txt".format(vstr)),
+#    join(gchp_vs_gchp_plotsdir, "{}_sig_diffs_500hpa.txt".format(vstr)),
+#    join(gchp_vs_gchp_plotsdir, "{}_sig_diffs_zonalmean.txt".format(vstr)),
+#    join(gchp_vs_gchp_plotsdir, "{}_sig_diffs_emissions.txt").format(vstr)]
 ##############################################################################
 
 # ======================================================================
@@ -236,19 +236,19 @@ if gcc_vs_gcc:
         # --------------------------------------------------------------
         # GCC vs GCC Concentration plots
         # --------------------------------------------------------------
-        title = '\n%%% Creating GCC vs. GCC {} concentration plots %%%'.format(
+        title = "\n%%% Creating GCC vs. GCC {} concentration plots %%%".format(
             bmk_type)
         print(title)
    
         # File lists for emissions data (seasonal)
-        collection = 'SpeciesConc'
+        collection = "SpeciesConc"
         gcc_vs_gcc_refspc = get_filepaths(gcc_vs_gcc_refdir, collection,
                                           bmk_seasons, is_gcc=True)
         gcc_vs_gcc_devspc = get_filepaths(gcc_vs_gcc_devdir, collection,
                                           bmk_seasons, is_gcc=True)
 
         # Only plot concentration categories for TransportTracers
-        restrict_cats = ['RnPbBeTracers', 'PassiveTracers']
+        restrict_cats = ["RnPbBeTracers", "PassiveTracers"]
 
         # Create seasonal concentration plots
         for s in range(bmk_nseasons):
@@ -270,10 +270,10 @@ if gcc_vs_gcc:
         # --------------------------------------------------------------
         # GCC vs GCC wet deposition
         # --------------------------------------------------------------
-        print('\n%%% Creating GCC vs. GCC wet deposition plots %%%')
+        print("\n%%% Creating GCC vs. GCC wet deposition plots %%%")
                 
         # Loop over wet deposition collections
-        collection_list = ['WetLossConv', 'WetLossLS']
+        collection_list = ["WetLossConv", "WetLossLS"]
         for collection in collection_list:
             gcc_vs_gcc_refwd = get_filepaths(gcc_vs_gcc_refdir, collection, 
                                              bmk_seasons, is_gcc=True)
@@ -300,15 +300,15 @@ if gcc_vs_gcc:
         # --------------------------------------------------------------
         # GCC vs GCC budgets tables
         # --------------------------------------------------------------
-        title = '\n%%% Creating GCC vs. GCC {} budgets %%%'.format(bmk_type)
+        title = "\n%%% Creating GCC vs. GCC {} budgets %%%".format(bmk_type)
         print(title)
 
         # Budgets of Radionuclide species
-        plot_dir = join(gcc_vs_gcc_plotsdir, 'Tables')
+        plot_dir = join(gcc_vs_gcc_plotsdir, "Tables")
         ttbdg.transport_tracers_budgets(maindir,
                                         gcc_dev_version,
-                                        plot_dir,
                                         bmk_year,
+                                        dst=plot_dir,
                                         overwrite=True)
 
         # Change in mass of species after each operation
@@ -333,9 +333,9 @@ if gcc_vs_gcc:
                    "SHEmis90dayTracer"]
         opbdg.make_operations_budget_table(gcc_dev_version,
                                            gcc_vs_gcc_devbgt,
-                                           plot_dir,
                                            bmk_type,
                                            label,
+                                           dst=plot_dir,
                                            interval=sec_per_yr,
                                            species=species,
                                            overwrite=True)
@@ -345,20 +345,20 @@ if gcc_vs_gcc:
         # GCC Strat-Trop Exchange
         # --------------------------------------------------------------
         title = \
-        '\n%%% Creating GCC vs. GCC {} Strat-Trop Exchange table %%%'.format(
+        "\n%%% Creating GCC vs. GCC {} Strat-Trop Exchange table %%%".format(
             bmk_type)
         print(title)
         
         # Strat-trop exchange of radionuclide species
         gcc_vs_gcc_devflx = get_filepaths(gcc_vs_gcc_devdir, "AdvFluxVert",
                                           bmk_months, is_gcc=True)
-        plot_dir = join(gcc_vs_gcc_plotsdir, 'Tables')
+        plot_dir = join(gcc_vs_gcc_plotsdir, "Tables")
         species = ["Pb210", "Be7", "Be10"]
         ste.make_benchmark_ste_table(gcc_dev_version,
                                      gcc_vs_gcc_devflx,
-                                     plot_dir,
                                      bmk_year,
-                                     bmk_type,
+                                     dst=plot_dir,
+                                     bmk_type=bmk_type,
                                      species=species,
                                      overwrite=True)
     
@@ -373,7 +373,7 @@ if gcc_vs_gcc:
 #    if plot_conc:
 #        # Concentration plots
 #        # (includes lumped species and separates by category)
-#        print('\n%%% Creating GCHP vs. GCC concentration plots %%%')
+#        print("\n%%% Creating GCHP vs. GCC concentration plots %%%")
 #        bmk.make_benchmark_conc_plots(gchp_vs_gcc_refspc,
 #                                      gchp_vs_gcc_refstr,
 #                                      gchp_vs_gcc_devspc,
@@ -384,13 +384,13 @@ if gcc_vs_gcc:
 #
 #    if plot_emis:
 #        # Emissions plots
-#        print('\n%%% Creating GCHP vs. GCC emissions plots %%%')
+#        print("\n%%% Creating GCHP vs. GCC emissions plots %%%")
 #        bmk.make_benchmark_emis_plots(gchp_vs_gcc_refhco,
 #                                      gchp_vs_gcc_refstr,
 #                                      gchp_vs_gcc_devhco,
 #                                      gchp_vs_gcc_devstr,
 #                                      dst=gchp_vs_gcc_plotsdir,
-#                                      plot_by_benchmark_cat=True,
+#                                      plot_by_spc_cat=True,
 #                                      plot_by_hco_cat=True,
 #                                      overwrite=True,
 #                                      flip_dev=True,
@@ -398,7 +398,7 @@ if gcc_vs_gcc:
 #
 #    if emis_table:
 #        # Tables of emissions and inventory totals
-#        print('\n%%% Creating GCHP vs. GCC emissions and inventory tables %%%')
+#        print("\n%%% Creating GCHP vs. GCC emissions and inventory tables %%%")
 #        gchp_vs_gcc_reflist = [gchp_vs_gcc_refhco]
 #        gchp_vs_gcc_devlist = [gchp_vs_gcc_devhco, gchp_vs_gcc_devmet]
 #        bmk.make_benchmark_emis_tables(gchp_vs_gcc_reflist,
@@ -410,7 +410,7 @@ if gcc_vs_gcc:
 #
 #    if plot_jvalues:
 #        # Local noon J-values plots
-#        print('\n%%% Creating GCHP vs. GCC J-value plots %%%')
+#        print("\n%%% Creating GCHP vs. GCC J-value plots %%%")
 #        bmk.make_benchmark_jvalue_plots(gchp_vs_gcc_refjv,
 #                                        gchp_vs_gcc_refstr,
 #                                        gchp_vs_gcc_devjv,
@@ -421,7 +421,7 @@ if gcc_vs_gcc:
 #
 #    if plot_aod:
 #        # Column AOD plots
-#        print('\n%%% Creating GCHP vs. GCC column AOD plots %%%')
+#        print("\n%%% Creating GCHP vs. GCC column AOD plots %%%")
 #        bmk.make_benchmark_aod_plots(gchp_vs_gcc_refaod,
 #                                     gchp_vs_gcc_refstr,
 #                                     gchp_vs_gcc_devaod,
@@ -433,7 +433,7 @@ if gcc_vs_gcc:
 ## Under development, leave commented out for now (bmy, 9/5/19)
 ##    if mass_table:
 ##        # Global mass tables
-##        print('\n%%% Creating GCHP vs. GCC global mass tables %%%')
+##        print("\n%%% Creating GCHP vs. GCC global mass tables %%%")
 ##        gchp_vs_gcc_reflist = [gchp_vs_gcc_refrst]
 ##        gchp_vs_gcc_devlist = [gchp_vs_gcc_devrst, gchp_vs_gcc_devmetinst]
 ##        bmk.make_benchmark_mass_tables(gchp_vs_gcc_reflist,
@@ -445,7 +445,7 @@ if gcc_vs_gcc:
 #        
 #    if OH_metrics:
 #        # Global mean OH, MCF Lifetime, CH4 Lifetime
-#        print('\n%%% Creating GCHP vs. GCC OH metrics %%%')
+#        print("\n%%% Creating GCHP vs. GCC OH metrics %%%")
 #        gchp_vs_gcc_reflist = [gchp_vs_gcc_refcac, gchp_vs_gcc_refmet]
 #        gchp_vs_gcc_devlist = [gchp_vs_gcc_devcac, gchp_vs_gcc_devmet]
 #        bmk.make_benchmark_oh_metrics(gchp_vs_gcc_reflist,
@@ -464,7 +464,7 @@ if gcc_vs_gcc:
 #    if plot_conc:
 #        # Concentration plots
 #        # (includes lumped species and separates by category)
-#        print('\n%%% Creating GCHP vs. GCHP concentration plots %%%')
+#        print("\n%%% Creating GCHP vs. GCHP concentration plots %%%")
 #        bmk.make_benchmark_conc_plots(gchp_vs_gchp_refspc,
 #                                      gchp_vs_gchp_refstr,
 #                                      gchp_vs_gchp_devspc,
@@ -475,13 +475,13 @@ if gcc_vs_gcc:
 #
 #    if plot_emis:
 #        # Emissions plots
-#        print('\n%%% Creating GCHP vs. GCHP emissions plots %%%')
+#        print("\n%%% Creating GCHP vs. GCHP emissions plots %%%")
 #        bmk.make_benchmark_emis_plots(gchp_vs_gchp_refhco,
 #                                      gchp_vs_gchp_refstr,
 #                                      gchp_vs_gchp_devhco,
 #                                      gchp_vs_gchp_devstr,
 #                                      dst=gchp_vs_gchp_plotsdir,
-#                                      plot_by_benchmark_cat=True,
+#                                      plot_by_spc_cat=True,
 #                                      plot_by_hco_cat=True,
 #                                      overwrite=True,
 #                                      flip_ref=True,
@@ -490,7 +490,7 @@ if gcc_vs_gcc:
 #
 #    if emis_table:
 #        # Tables of emissions and inventory totals
-#        print('\n%%% Creating GCHP vs. GCHP emissions and inventory tables %%%')
+#        print("\n%%% Creating GCHP vs. GCHP emissions and inventory tables %%%")
 #        gchp_vs_gchp_reflist = [gchp_vs_gchp_refhco, gchp_vs_gchp_refmet]
 #        gchp_vs_gchp_devlist = [gchp_vs_gchp_devhco, gchp_vs_gchp_devmet]
 #        bmk.make_benchmark_emis_tables(gchp_vs_gchp_reflist,
@@ -502,7 +502,7 @@ if gcc_vs_gcc:
 #
 #    if plot_jvalues:
 #        # Local noon J-values plots
-#        print('\n%%% Creating GCHP vs. GCHP J-value plots %%%')
+#        print("\n%%% Creating GCHP vs. GCHP J-value plots %%%")
 #        bmk.make_benchmark_jvalue_plots(gchp_vs_gchp_refjv,
 #                                        gchp_vs_gchp_refstr,
 #                                        gchp_vs_gchp_devjv,
@@ -513,7 +513,7 @@ if gcc_vs_gcc:
 #
 #    if plot_aod:
 #        # Column AOD plots
-#        print('\n%%% Creating GCHP vs. GCHP column AOD plots %%%')
+#        print("\n%%% Creating GCHP vs. GCHP column AOD plots %%%")
 #        bmk.make_benchmark_aod_plots(gchp_vs_gchp_refaod,
 #                                     gchp_vs_gchp_refstr,
 #                                     gchp_vs_gchp_devaod,
@@ -525,7 +525,7 @@ if gcc_vs_gcc:
 ## Under development, leave commented out for now (bmy, 9/5/19)
 ##    if mass_table:
 ##        # Global mass tables
-##        print('\n%%% Creating GCHP vs. GCHP global mass tables %%%')
+##        print("\n%%% Creating GCHP vs. GCHP global mass tables %%%")
 ##        gchp_vs_gchp_reflist = [gchp_vs_gchp_refrst, gchp_vs_gchp_refmetinst]
 ##        gchp_vs_gchp_devlist = [gchp_vs_gchp_devrst, gchp_vs_gchp_devmetinst]
 ##        bmk.make_benchmark_mass_tables(gchp_vs_gchp_reflist,
@@ -537,7 +537,7 @@ if gcc_vs_gcc:
 #
 #    if OH_metrics:
 #        # Global mean OH, MCF Lifetime, CH4 Lifetime
-#        print('\n%%% Creating GCHP vs. GCHP OH metrics %%%')
+#        print("\n%%% Creating GCHP vs. GCHP OH metrics %%%")
 #        gchp_vs_gcc_reflist = [gchp_vs_gchp_refcac, gchp_vs_gchp_refmet]
 #        gchp_vs_gcc_devlist = [gchp_vs_gchp_devcac, gchp_vs_gchp_devmet]
 #        bmk.make_benchmark_oh_metrics(gchp_vs_gchp_reflist,
@@ -555,14 +555,14 @@ if gcc_vs_gcc:
 #
 #    # NOTE: This can be expanded to differences beyond species
 #    # concentrations by following how this is done for conc plots.
-#    print('%\n%% Creating GCHP vs. GCC diff-of-diffs concentration plots %%%')
+#    print("%\n%% Creating GCHP vs. GCC diff-of-diffs concentration plots %%%")
 #
 #    # Get a list of variables that GCPy should not read
 #    skip_vars = skip_these_vars
 #
 #    # Target output files
-#    diff_of_diffs_refspc = './gcc_diffs_spc.nc4'
-#    diff_of_diffs_devspc = './gchp_diffs_spc.nc4'
+#    diff_of_diffs_refspc = "./gcc_diffs_spc.nc4"
+#    diff_of_diffs_devspc = "./gchp_diffs_spc.nc4"
 #
 #    # Create a ref file that contains GCC differences
 #    gcc_ref  = xr.open_dataset(gcc_vs_gcc_refspc, drop_variables=skip_vars)
@@ -581,18 +581,18 @@ if gcc_vs_gcc:
 #    gchp_dev = xr.open_dataset(gchp_vs_gchp_devspc, drop_variables=skip_vars)
 #    refdims = gchp_ref.dims
 #    devdims = gchp_dev.dims
-#    if 'lat' in refdims and 'Xdim' in devdims:
+#    if "lat" in refdims and "Xdim" in devdims:
 #        gchp_ref_newdimnames = gchp_dev.copy()
 #        for v in gchp_dev.data_vars.keys():
-#            if 'Xdim' in gchp_dev[v].dims:
+#            if "Xdim" in gchp_dev[v].dims:
 #                gchp_ref_newdimnames[v].values = gchp_ref[v].values.reshape(
 #                    gchp_dev[v].values.shape)
-#                # NOTE: the reverse conversion is gchp_dev[v].stack(lat=('nf','Ydim')).transpose('time','lev','lat','Xdim').values
+#                # NOTE: the reverse conversion is gchp_dev[v].stack(lat=("nf","Ydim")).transpose("time","lev","lat","Xdim").values
 #        gchp_ref = gchp_ref_newdimnames.copy()
 #    with xr.set_options(keep_attrs=True):
 #        gchp_diffs = gchp_dev.copy()
 #        for v in gchp_dev.data_vars.keys():
-#            if 'Xdim' in gchp_dev[v].dims or 'lat' in gchp_dev[v].dims:
+#            if "Xdim" in gchp_dev[v].dims or "lat" in gchp_dev[v].dims:
 #                gchp_diffs[v] = gchp_dev[v] - gchp_ref[v]
 #                # NOTE: The gchp_diffs Dataset is created without variable
 #                # attributes; we have to reattach them