+++ CHANGELOG.md
@@ -9,13 +9,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Added strat/trop exchange fluxes to 1-year benchmark output (gcpy/ste_flux.py)
 - Added operations budgets to 1-year benchmark output (gcpy/budget_ops.py)
-- Added seasonal mass table output for 1-year benchmak (FullChem).
+- Added seasonal mass table output for 1-year FullChemBenchmark.
+- Added mean OH from log files for 1-year FullChemBenchmark (GC-Classic only).
 
 ### Changed
 - Significant difference files are now written out to the Plots/Sig_Diffs folder for the 1-year benchmarks.
 - Updated file names for Pb/Be budget tables in gcpy/budgets_tt.py.
 - Created separate driver routines for 1-year FullChem and TransportTracers benchmarks
 - Useless warnings when creating benchmark output should now be suppressed
+- Can now create benchmark plots in a single file instead of by category.
 
 ### Deprecated